OBPIH-7001 Move parseCsvBooleanField method to CsvUtils file
